migrate cli to internal package (#301)
