+++ internal/cli/account_import.go
@@ -1,11 +1,11 @@
-package main
+package cli
 
 import (
 	"fmt"
 
 	"github.com/ethereum/go-ethereum/cmd/utils"
-	"github.com/ethereum/go-ethereum/command/flagset"
 	"github.com/ethereum/go-ethereum/crypto"
+	"github.com/ethereum/go-ethereum/internal/cli/flagset"
 )
 
 type AccountImportCommand struct {
