+++ internal/cli/account_list.go
@@ -1,10 +1,10 @@
-package main
+package cli
 
 import (
 	"fmt"
 
 	"github.com/ethereum/go-ethereum/accounts"
-	"github.com/ethereum/go-ethereum/command/flagset"
+	"github.com/ethereum/go-ethereum/internal/cli/flagset"
 )
 
 type AccountListCommand struct {
