+++ internal/cli/server/chains/allocs/mainnet.json
