+++ internal/cli/server/config_legacy_test.go
