+++ internal/cli/server/proto/server.proto
