+++ internal/cli/server/testdata/simple.hcl
