+++ internal/cli/server/testdata/simple.json
