+++ internal/cli/version.go
@@ -1,4 +1,4 @@
-package main
+package cli
 
 import (
 	"github.com/ethereum/go-ethereum/params"
